Created a basic platform layer to initalize OpenGL
Many a thing happened here.  First off GLEW v2.1.0 was added as a static library.  In accordance
with this build.bat was updated to compile correctly with the new library.  Several windows lib files
were also added to the build.  In terms of code a debug OpenGL debug context is properly created using
the extended context creation methods.  A window is opened and messages are processed properly.  The
rendering function merely displays the clear color to ensure OpenGL is working correctly.  The point
where rendering code should be is clearly labeled.  I think that covers it all.  Not sure.  It's late...