Changed to rendering a wireframe of the mesh and added rotation to prove it's a cube
The vertex shader now applies a rotation matrix based on a unfiorm angle passed in during
GLRenderMesh.  This float is computed by taking the system time in milliseconds, clamping it to
a value between 0 and 359, and multiplying it by 0.1.  Also started validating GL programs when they're
created and refactored GLRender to GLRenderMesh since that's essentially what it's doing.  In order to make the
wireframe more visible the static color of the fragment shader has been changed to black.