Dirty renderer set up, read description
In this implementation, there is a 2D renderer that you may submit a
renderable to. There are also entities that may exist relative to screen
space but do not get drawn. Now by ignoring all other files aside from
GameLoop (to be renamed, it isn't a loop at all it's the interface for
the renderer) we may fully draw simple sprites and modify shader values.

Later the plan is the submit verticies and indicies only once and use
geometry shaders to manipulate our renderables.

As it is now, we have a projection, model, and view matrix setup as well
as a good interface for them. This will expand of course.